+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to `laravel-i18n` will be documented in this file.
 
+## 0.1.6 - 2018-01-28
+* Improved route registration logic
+* On routes you can call `forceRegionPrefix()` to force region when the `default` setting is set.
+
 ## 0.1.5 - 2018-01-26
 * I18n: manually set the current region with `setRegion($region)`
 * I18n: added `getDefaultRegion()` method